Hide VLM files and API (#951)
