+++ src/cpp/src/visual_language/clip.hpp
