+++ src/cpp/src/visual_language/vlm_model_type.hpp
