+++ src/python/py_vlm_pipeline.cpp
@@ -7,7 +7,7 @@
 #include <pybind11/stl.h>
 #include <pybind11/stl_bind.h>
 #include <pybind11/functional.h>
-#include "openvino/genai/vlm_pipeline.hpp"
+#include "openvino/genai/visual_language/pipeline.hpp"
 #include "../cpp/src/tokenizers_path.hpp"
 #include "./utils.hpp"
